added exporting licenses, synchronization of licenses, configuration documentation
